#refactor : reorganize components & css directory
