+++ components/header/header.css
@@ -1,0 +1,23 @@
+header {
+    height: 64px;
+    padding: 0px 80px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+header > span {
+    display: flex;
+    align-items: center;
+}
+
+#sort_button {
+    margin-left: 16px;
+    box-sizing:content-box;
+    padding: 3px 8px;
+    display: flex;
+    align-items: center;
+    span {
+        padding: 0 4px;
+    }
+}
