+++ index.html
@@ -2,18 +2,19 @@
 <html lang="en">
 
 <head>
-    <link rel="stylesheet" href="public/globals.css">
-    <link rel="stylesheet" href="public/header.css">
-    <link rel="stylesheet" href="public/main.css">
-    <link rel="stylesheet" href="public/form.css">
-    <link rel="stylesheet" href="public/modal.css">
+    <link rel="stylesheet" href="globals.css">
+    <link rel="stylesheet" href="./components/main/main.css">
+    <link rel="stylesheet" href="./components/header/header.css">
+    <link rel="stylesheet" href="./components/form/form.css">
+    <link rel="stylesheet" href="./components/modal/modal.css">
+    <link rel="stylesheet" href="./components/log/log.css">
+    <link rel="stylesheet" href="./components/column/column.css">
+    <link rel="stylesheet" href="./components/task/task.css">
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
 </head>
-
 <body class="surface-alt">
 </body>
 <script type="module" src="/script/index.js"></script>
-
 </html>
