--- public/header.css
@@ -1,7 +1,0 @@
-header {
-    height: 64px;
-    padding: 0px 80px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
