--- public/main.css
@@ -1,251 +1,0 @@
-main {
-    height: calc(100vh - 64px);
-    padding: 32px 80px;
-}
-
-header > span {
-    display: flex;
-    align-items: center;
-}
-
-#sort_button {
-    margin-left: 16px;
-    box-sizing:content-box;
-    padding: 3px 8px;
-    display: flex;
-    align-items: center;
-    span {
-        padding: 0 4px;
-    }
-}
-
-#column_container {
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-    list-style-type: none;
-}
-
-
-.column {
-    width: 320px;
-    min-width: 320px;
-    margin-right: 32px;
-
-    &:last-of-type {
-        margin: 0;
-    }
-}
-
-.column>ol {
-    height: 100%;
-    list-style-type: none;
-}
-
-.column_header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 16px;
-
-    span {
-        display: flex;
-
-        button {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-        }
-    }
-
-    input {
-        width: 100%;
-        outline: none;
-        border-radius: 6px;
-        padding: 0 8px;
-        height: 24px;
-    }
-}
-
-.column_title {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.column_button_container {
-    margin-left: 8px;
-
-    >button {
-        &:first-of-type {
-            &:hover {
-                filter: invert(34%) sepia(33%) saturate(6685%) hue-rotate(200deg) brightness(102%) contrast(106%);
-            }
-
-        }
-
-        &:last-of-type {
-            &:hover {
-                filter: invert(35%) sepia(50%) saturate(6582%) hue-rotate(343deg) brightness(110%) contrast(99%);
-            }
-        }
-    }
-}
-
-.column_task_counter {
-    margin-left: 8px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-width: 24px;
-    padding: 3.5px 4px;
-    text-align: center;
-
-}
-
-
-
-.card {
-    display: flex;
-    flex-direction: row;
-    margin-top: 10px;
-    padding: 16px;
-}
-
-.card_text_container {
-    flex-grow: 1;
-    word-break: break-all;
-
-    h4 {
-        margin-bottom: 8px;
-    }
-}
-
-.card_button_container {
-    align-items: flex-end;
-    display: flex;
-    flex-direction: column;
-    padding: 3px;
-}
-
-#log_layer {
-    position: absolute;
-    top: 64px;
-    right: 62px;
-    width: 320px;
-    padding: 8px;
-}
-
-#log_layer_footer {
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    padding: 0 8px;
-    
-    > button {
-        display: block;
-        margin: 8px ;
-        padding: 0 4px;
-        text-align: center;
-    }
-}
-
-#log_layer_header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    flex-direction: row;
-    padding: 8px;
-    padding-left: 16px;
-
-    button {
-        display: flex;
-        align-items: center;
-        padding: 8px;
-    }
-
-}
-
-#log_layer_list {
-    overflow: auto;
-    max-height: 600px;
-    margin: 8px 0;
-}
-
-#log_layer_list>li {
-    padding: 16px;
-    display: flex;
-    flex-direction: row;
-    border-bottom: 1px solid;
-    border-color: var(--grayscale_200);
-
-    >:first-child {
-
-        margin-right: 16px;
-
-        .log_profile {
-            width: 40px;
-            height: 40px;
-            border-radius: 50%;
-        }
-    }
-
-    >:last-child {
-        width: 100%;
-
-        p:nth-of-type(2) {
-            margin-top: 8px;
-            margin-bottom: 8px;
-
-            span {
-                color: var(--grayscale_700);
-                font-size: 14px;
-                line-height: auto;
-                font-weight: 700;
-            }
-        }
-    }
-
-    &:last-of-type {
-        border: none;
-    }
-}
-
-#log_layer_empty {
-    padding: 16px;
-
-    p {
-        width: 100%;
-        text-align: center;
-    }
-}
-
-#fab_container {
-    position: absolute;
-    right: 48px;
-    bottom: 48px;    
-}
-
-.fab_button {
-    padding: 12px;
-    margin-bottom: 16px;
-    
-    &:last-of-type {
-        margin: 0px;
-    }
-}
-
-
-.vibration {
-    animation: vibration .1s infinite;
-  }
-  
-  @keyframes vibration {
-    from {
-      transform: translateX(2px);
-    }
-    to {
-        transform: translateX(-2px);
-    }
-  }
